Add .spec file for pyinstaller to produce warning-free builds
Fair warning: I don't know much about Pyinstaller. I'm committing the
.spec file here so we don't lose the knowledge about how to create
a warning-free binary.

This all started after running:

$ pyinstaller app.py -F

It creates a binary called app in the dist directory. And the binary
works as expected, but always shows this warning when it launches:

[56539] WARNING: file already exists but should not: /var/folders/h1/y9t7drkj15gbnb1mjs1q7pxc0000gp/T/_MEI2FeUjl/pyarrow/lib.cpython-38-darwin.so

Annoying...

I found a stack overflow question[1] about removing this kind of error.
It recommended removing all the binaries that match the offending name.
In my case, that's "lib.cptyhon-38-darwin.so".

But when I tried the suggested loop it deleted *all* copies of the .so,
which broke the binary. It crashed with an error that lib.cpython-38-darwin.so
was needed but not found.

Oops...

The ultimate solution I borrowed from Pyinstaller#4561[2]. This way it
only deletes the _first_ matching file. After that, it breaks out of the
loop before deleting any other matching files. Presumably, if there were
more than two copies of the .so file, this loop could be adjusted to
delete all but the _last_ one.

[1]: https://stackoverflow.com/questions/66069360/pyinstaller-onefile-warning-file-already-exists-but-should-not
[2]: https://github.com/pyinstaller/pyinstaller/issues/4561#issuecomment-836155176