Remove custom use command.
dbsqlcli was automatically capturing instances of "use xxxxx" and failing
silently to connect. For example "use catalog main" would try to recreate
the dbsql connection with schema="catalog main", which fails. But this
didn't raise any exceptions.

Removing this custom command allows queries to pass unintercepted to the
sql warehouse or all-purpose cluster.

Signed-off-by: Jesse Whitehouse <jesse.whitehouse@databricks.com>